[PiezoSpeaker] Changed namespace to Netduino.Foundation.Audio and moved sample physical location into Samples folder
