+++ Source/Netduino.Foundation.Core.Samples/PiezoSample/NoteFrequencies.cs
@@ -1,4 +1,4 @@
-namespace Netduino.Foundation.Piezos
+namespace PiezoSample
 {
     public static class NoteFrequencies
     {
