+++ Source/Netduino.Foundation.Core.Samples/PiezoSample/Properties/AssemblyInfo.cs
